Add check for one person masonries
